전체 컴포넌트화 리팩토링 작업 및 css 1차 수정
