+++ src/App.module.css
@@ -1,81 +1,4 @@
-header {
-  /* background-color: pink; */
-  width: 100%;
-  height: 7rem;
-  display: flex;
-  justify-content: space-between;
-  border-bottom: 1px solid var(--border-gray);
-  margin: 0 auto;
-}
-
-.header {
-  display: flex;
-}
-
-.header > .header_left {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 20rem;
-}
-
-.header > .header_left > .header_logo {
-  width: 4rem;
-  height: 4rem;
-}
-.header > .header_left > .logo_title {
-  color: var(--login-normal-blue);
-  font-size: 2.5rem;
-  font-weight: 700;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 0.8rem;
-}
-
-.header_nav {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  justify-content: space-between;
-  margin-left: 4.7rem;
-}
-
-.header_nav > .header_nav__menu {
-  display: flex;
-}
-
-.header_nav > .header_nav__menu > li {
-  margin-right: 4.7rem;
-  font-size: 1.8rem;
-  font-weight: 700;
-}
-.header_nav > .header_nav__menu > li:hover {
-  cursor: pointer;
-}
-.header_nav__menu > .used_market:active {
-  color: var(--login-normal-blue);
-}
-
-.header_right {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.header_right > .header_right__button {
-  width: 8.8rem;
-  height: 4.2rem;
-  font-size: 1.6rem;
-  font-weight: 600;
-  color: var(--login-normal-font-color);
-  background-color: var(--login-normal-blue);
-  border-radius: 0.8rem;
-  border: none;
-  margin-right: 20rem;
-}
-
-.main_product {
+/* .main_product {
   width: 100%;
 }
 
@@ -153,7 +76,9 @@ header {
 .main_product_detail > .product_description {
   font-size: 1.4rem;
   width: 14rem;
-  white-space: nowrap; /* 텍스트가 한 줄로만 표시되도록 설정 */
-  overflow: hidden; /* 넘치는 텍스트는 숨김 */
-  text-overflow: ellipsis; /* 넘치는 텍스트는 "..."로 대체 */
+} */
+
+main {
+  width: 120rem;
+  margin: 0 auto;
 }
