+++ src/components/AllProduct.module.css
@@ -1,0 +1,96 @@
+.main_allproduct_container {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 4rem;
+  width: 100%;
+}
+
+.allproduct_header_title {
+  font-size: 2rem;
+  font-weight: 700;
+  line-height: 2.8rem;
+  color: var(--gray-900);
+  margin-bottom: 1.6rem;
+}
+
+.all_product_header_right {
+  width: 62rem;
+  display: flex;
+  justify-content: space-between;
+}
+
+.search_product {
+  width: 32.5rem;
+  height: 4.2rem;
+  border: none;
+  border-radius: 1.2rem;
+  padding-left: 1rem;
+  background-color: var(--gray-100);
+}
+
+.search_product::placeholder {
+  color: var(--gray-400);
+}
+
+.all_product_additem {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-decoration: none;
+  border-radius: 8px;
+  background-color: var(--login-normal-blue);
+  color: var(--login-normal-font-color);
+  font-size: 1.6rem;
+  font-weight: 600;
+  width: 13.3rem;
+  height: 4.2rem;
+}
+
+.orderBy {
+  border: 0.1rem solid var(--gray-200);
+  border-radius: 1.2rem;
+  width: 13rem;
+  height: 4.2rem;
+  padding-left: 1rem;
+}
+
+.main_all_product {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  height: 60rem;
+  gap: 1rem;
+}
+
+.product_name,
+.product_price {
+  color: var(--gray-800);
+}
+
+.product_name {
+  font-size: 1.4rem;
+  font-weight: 500;
+  line-height: 1.6rem;
+  margin-top: 1.6rem;
+  margin-bottom: 0.6rem;
+}
+
+.product_price {
+  font-size: 1.6rem;
+  font-weight: 700;
+  line-height: 1.9rem;
+  margin-bottom: 0.6rem;
+}
+
+.product_favorite {
+  color: var(--gray-600);
+  font-size: 1.2rem;
+  font-weight: 500;
+}
+
+.pageButton {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 5rem;
+}
