+++ src/components/BestProduct.module.css
@@ -1,0 +1,26 @@
+.main_product {
+  width: 100%;
+}
+
+.main_product > .main_product_container {
+  width: 120.1rem;
+  height: 115.4rem;
+  margin: 0 auto;
+  margin-top: 2.4rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.main_product_bestproduct > .main_product_bestproduct_title {
+  font-size: 2rem;
+  font-weight: 700;
+  line-height: 2.8rem;
+  color: var(--gray-900);
+  margin-bottom: 1.6rem;
+}
+
+.main_best_product {
+  display: flex;
+  gap: 1rem;
+}
