+++ src/global.css
@@ -4,7 +4,9 @@
   --login-normal-font-color: #ffffff;
   --gray-900: #111827;
   --gray-800: #1f2937;
+  --gray-600: #4b5563;
   --gray-400: #9ca3af;
+  --gray-200: #e5e7eb;
   --gray-100: #f3f4f6;
 }
